Rename trace_processor -> trace_reader
