+++ cloud/blockstore/libs/diagnostics/config.h
@@ -4,7 +4,7 @@
 
 #include <cloud/blockstore/config/diagnostics.pb.h>
 
-#include <cloud/storage/core/libs/diagnostics/trace_processor.h>
+#include <cloud/storage/core/libs/diagnostics/trace_reader.h>
 
 #include <util/generic/string.h>
 #include <util/stream/output.h>
